requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
google-generativeai>=0.4.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy<2
requests==2.31.0
python-dotenv==1.0.1
